Dot15d4 update
- Chose between Zigbee and SixLowPan using config
- Better FCS support in Dot15d4
- Move Zigbee support to a different layer
- Minor fixes
