+++ scapy/data.py
@@ -88,11 +88,13 @@ DLT_IEEE802_11_RADIO = 127
 DLT_LINUX_IRDA = 144
 DLT_IEEE802_11_RADIO_AVS = 163
 DLT_BLUETOOTH_HCI_H4 = 187
+DLT_IEEE802_15_4_WITHFCS = 195
 DLT_BLUETOOTH_HCI_H4_WITH_PHDR = 201  # TODO: IMPLEMENT (currently unimplemented)  # noqa: E501
 DLT_PPI = 192
 DLT_CAN_SOCKETCAN = 227
 DLT_IPV4 = 228
 DLT_IPV6 = 229
+DLT_IEEE802_15_4_NOFCS = 230
 DLT_BLUETOOTH_LE_LL = 251
 
 # From net/ipv6.h on Linux (+ Additions)
